+++ src/hooks/useFormInput/index.ts
@@ -1,0 +1,2 @@
+export { useFormInput } from './useFormInput';
+export { FormInputHookResult, FormOutputObject, FormInputObject } from './types';
